Renamed GuiMacroEditor2 to GuiMacroEditor
