+++ src/client/java/tools/redstone/redstonetools/malilib/widget/macro/WidgetMacroEntry.java
@@ -8,7 +8,7 @@ import fi.dy.masa.malilib.gui.button.IButtonActionListener;
 import fi.dy.masa.malilib.gui.widgets.WidgetListEntryBase;
 import fi.dy.masa.malilib.util.StringUtils;
 import net.minecraft.text.Text;
-import tools.redstone.redstonetools.malilib.GuiMacroEditor2;
+import tools.redstone.redstonetools.malilib.GuiMacroEditor;
 import tools.redstone.redstonetools.malilib.config.MacroManager;
 
 public class WidgetMacroEntry extends WidgetListEntryBase<MacroBase> {
@@ -56,7 +56,7 @@ public class WidgetMacroEntry extends WidgetListEntryBase<MacroBase> {
 		@Override
 		public void actionPerformedWithButton(ButtonBase button, int mouseButton) {
 			if (this.type == Type.CONFIGURE) {
-				GuiMacroEditor2 gui = new GuiMacroEditor2(Text.of(this.widget.macro.name), this.widget.macro, this.widget.parent.parent);
+				GuiMacroEditor gui = new GuiMacroEditor(Text.of(this.widget.macro.name), this.widget.macro, this.widget.parent.parent);
 				GuiBase.openGui(gui);
 				this.widget.parent.refreshEntries();
 			} else if (this.type == Type.REMOVE) {
